app/src/backend/requirements.txt
pytest-asyncio>=0.24.0
python-dotenv==1.0.0
python-json-logger==3.2.1
python-Levenshtein==0.26.1
python-multipart==0.0.18
python_docx==1.1.2
rich==13.6.0
sqlalchemy==2.0.21
uvicorn==0.23.2
